Created decorators for skip_if locale functions
